+++ .eslintrc.json
@@ -2,5 +2,5 @@
   "extends": ["next/core-web-vitals", "next/typescript", "prettier"],
   "plugins": ["@typescript-eslint", "prettier"],
   "rules": { "@typescript-eslint/no-unused-vars": "error" },
-  "ignorePatterns": ["src/types/*"]
+  "ignorePatterns": ["src/types/unchecked/*"]
 }
