+++ global.d.ts
@@ -5,11 +5,7 @@ import type {
   ApiResponseSuccess as _ApiResponseSuccess,
   ApiResponseError as _ApiResponseError,
 } from '@/types/dto/ApiResponse';
-import type {
-  Empty as _Empty,
-  VoidifyReturn as _VoidifyReturn,
-  ReplaceValues as _ReplaceValues,
-} from '@/types/utility.types';
+import type { ReplaceValues as _ReplaceValues } from '@/types/utility.types';
 
 declare global {
   type SetState<T> = React.Dispatch<React.SetStateAction<T>>;
@@ -19,9 +15,6 @@ declare global {
   type ApiResponse<T> = _ApiResponse<T>;
   type ApiResponseSuccess<T> = _ApiResponseSuccess<T>;
   type ApiResponseError = _ApiResponseError;
-
-  type Empty = _Empty;
-  type VoidifyReturn<T> = _VoidifyReturn<T>;
   type ReplaceValues<T, R> = _ReplaceValues<T, R>;
   declare namespace NodeJS {
     interface ProcessEnv extends Record<EnvKey, string> {}
