--- src/types/dto/ReqCourseDetail.ts
@@ -1,4 +1,0 @@
-interface ReqCourseDetail extends ReqCourseRelated {
-  order: string;
-  page: number;
-}
