--- src/types/dto/ReqCourseRelated.ts
@@ -1,3 +1,0 @@
-interface ReqCourseRelated {
-  course_id: number;
-}
