--- src/types/notice/Notice.ts
@@ -1,4 +1,0 @@
-interface Notice<DateExpression = Date> extends NoticeMetadata<DateExpression> {
-  filename: string;
-  content: React.ReactNode;
-}
