--- src/types/notice/NoticeMetadata.ts
@@ -1,6 +1,0 @@
-interface NoticeMetadata<DateExpression> extends Record<string, unknown> {
-  title: string;
-  created_at: DateExpression;
-  writer: string;
-  tag: string[];
-}
