--- src/types/user/JWTDecoded.ts
@@ -1,7 +1,0 @@
-interface JWTDecoded {
-  profile_id: number;
-  email: string;
-  role: string;
-  iat: number;
-  exp: number;
-}
