--- src/types/user/MembershipData.ts
@@ -1,6 +1,0 @@
-interface MembershipData {
-  item: string;
-  day: number;
-  price: number;
-  iconLevel: 1 | 2 | 3;
-}
