--- src/types/user/UserProfile.ts
@@ -1,1 +1,0 @@
-type UserProfile = MyPageBasicInfo;
