+++ src/types/utility.types.ts
@@ -1,7 +1,3 @@
-export interface Empty {}
-
-export type VoidifyReturn<T> = T extends (...args: infer Args) => any ? (...args: Args) => void : never;
-
 export type ReplaceValues<T, K> = {
   [P in keyof T]: K;
 };
